Current state of progress:
Navbar is styled for a more modern look. Problems when styling collapsed toggler and normal navbar as of now. Currently WIP fix

Home Page is barely touched, and is only a launch page to make the user choose a more optimal approuch. Though PC is the only worthy.

PC Page is decently done. Several major things to improve.

Mobile Page is technically the PC Page, but just older and is redundant. Ignore it for now

Billetter Page is missing images due to relevancy to 007. Upon addition of images, further styling will happen.

Other Pages are just placeholder for the time being.